+++ Plugins/mapsplat/style_converter.py
@@ -17,7 +17,7 @@ Supported style properties:
 - Multiple symbol layers
 """
 
-__version__ = "0.5.11"
+__version__ = "0.6.0"
 
 import os
 from qgis.core import (